tests/wpt/web-platform-tests/scroll-animations/css/animation-timeline-none.html

<!DOCTYPE html>
<link rel="help" src="https://drafts.csswg.org/css-animations-2/#typedef-timeline-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<style>
  @keyframes expand {
    from { width: 100px; }
    to { width: 200px; }
  }

  .test, #ref {
    width: 0px;
    animation-name: expand;
    animation-duration: 10s;
    animation-play-state: paused;
  }

  #element_timeline_none {
    animation-timeline: none;
  }
  #element_unknown_timeline {
    animation-timeline: unknown_timeline;
  }

</style>
<div class=test id=element_timeline_none></div>
<div class=test id=element_unknown_timeline></div>
<div id=ref></div>
<script>
  promise_test(async (t) => {
    assert_equals(getComputedStyle(element_timeline_none).width, '0px');
    assert_equals(getComputedStyle(ref).width, '100px');
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_timeline_none).width, '0px');
    assert_equals(getComputedStyle(ref).width, '100px');
  }, 'Animation with animation-timeline:none has no effect value');

  promise_test(async (t) => {
    assert_equals(getComputedStyle(element_unknown_timeline).width, '0px');
    assert_equals(getComputedStyle(ref).width, '100px');
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_unknown_timeline).width, '0px');
    assert_equals(getComputedStyle(ref).width, '100px');
  }, 'Animation with unknown timeline name has no effect value');
</script>
